requirements.txt
reportlab
Django==2.2.5
pytz==2019.2
sqlparse==0.3.0
django-crispy-forms==1.7.2
django-datatables-view==1.15.1
mysqlclient==1.3.13
django-extensions==1.9.9
XlsxWriter
xhtml2pdf